+++ app/src/main/java/mihon/feature/upcoming/manga/UpcomingMangaViewModel.kt
@@ -53,6 +53,16 @@ class UpcomingMangaViewModel(
 
     private val dialog = MutableStateFlow<Dialog?>(null)
 
+    private val hasActiveFilters = getUpcomingItemPreferenceFlow()
+        .map { prefs ->
+            listOf(
+                prefs.filterIncludedCategories,
+                prefs.filterExcludedCategories,
+            )
+                .any { it.isNotEmpty() }
+        }
+        .distinctUntilChanged()
+
     private val upcoming = getUpcomingItemPreferenceFlow()
         .distinctUntilChanged()
         .flatMapLatest { prefs ->
@@ -61,27 +71,23 @@ class UpcomingMangaViewModel(
                 includedCategories = prefs.filterIncludedCategories,
             )
                 .distinctUntilChanged()
-                .map { items ->
-                    val upcomingItems = items.toUpcomingMangaUIModels()
-                    val hasFilters = prefs.filterExcludedCategories.isNotEmpty() ||
-                        prefs.filterIncludedCategories.isNotEmpty()
-                    Triple(upcomingItems, upcomingItems.toEvents(), hasFilters)
-                }
+                .map { items -> items.toUpcomingMangaUIModels() }
         }
         .flowOn(Dispatchers.IO)
-        .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5.seconds), Triple(emptyList(), emptyMap(), false))
+        .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5.seconds), emptyList())
 
     val state: StateFlow<State> = combine(
         upcoming,
         selectedYearMonth,
         dialog,
-    ) { (items, events, hasFilters), selectedYearMonth, dialog ->
+        hasActiveFilters,
+    ) { upcoming, selectedYearMonth, dialog, hasActiveFilters ->
         State(
             selectedYearMonth = selectedYearMonth,
-            items = items,
-            events = events,
-            headerIndexes = items.getHeaderIndexes(),
-            hasActiveFilters = hasFilters,
+            items = upcoming,
+            events = upcoming.toEvents(),
+            headerIndexes = upcoming.getHeaderIndexes(),
+            hasActiveFilters = hasActiveFilters,
             dialog = dialog,
         )
     }